Renamed Project to A4AA_Application. Moved the Survey Page to be accessed by the touch of a button, and main page is just two buttons that go to login page or survey page.
