+++ HelloWorld/HelloWorld/App.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <Application xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="HelloWorld.App">
+             x:Class="A4AA_Application.App">
 	<Application.Resources>
 
 		<!-- Application resource dictionary -->
