+++ HelloWorld/HelloWorld/Login/CreateAccountPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="HelloWorld.Login.CreateAccountPage">
+             x:Class="A4AA_Application.Login.CreateAccountPage">
     <ContentPage.Content>
         <StackLayout  HorizontalOptions ="CenterAndExpand" >
             <Label Text="Create Login!" FontSize="Large"/>
