+++ HelloWorld/HelloWorld/MainPage.xaml.cs
@@ -9,13 +9,14 @@ using Xamarin.Forms;
 using System.Reflection;
 using System.IO;
 
-namespace HelloWorld
+namespace A4AA_Application
 {
 	public partial class MainPage : ContentPage
 	{
-        public MainPage()
+		public MainPage()
 		{
 			InitializeComponent();
+			var layout = new StackLayout { Padding = new Thickness(5.0, 10.0) };
 
 			Button button = new Button//here
 			{
@@ -26,65 +27,31 @@ namespace HelloWorld
 				VerticalOptions = LayoutOptions.CenterAndExpand,
 				TextColor = Color.Red
 			};
-			button.Clicked += OnClicked;//to here
-			
+			button.Clicked += OnClicked1;//to here
+			layout.Children.Add(button);
 
-			var layout = new StackLayout { Padding = new Thickness(5.0, 10.0) };
+			Button button2 = new Button//here
+			{
+				Text = "Go to Survey Page!",
+				Font = Font.SystemFontOfSize(NamedSize.Large),
+				BorderWidth = 1,
+				HorizontalOptions = LayoutOptions.Center,
+				VerticalOptions = LayoutOptions.CenterAndExpand,
+				TextColor = Color.Red
+			};
+			button.Clicked += OnClicked2;//to here
+			layout.Children.Add(button2);
 
-			layout.Children.Add(button);//this
-
-			var assembly = typeof(App).GetTypeInfo().Assembly;
-            Stream stream = assembly.GetManifestResourceStream("HelloWorld.Droid.surveyQuestions.txt");
-            string text = "";
-            
-            using (var reader = new System.IO.StreamReader(stream))
-            {
-                text = reader.ReadToEnd();
-            }
-
-            string[] separatedLines = text.Split('\n');
-
-            foreach (string s in separatedLines)
-            {
-                //System.Diagnostics.Debug.WriteLine(s);
-                var label = new Label { FontSize = 20 };
-                var fs = new FormattedString();
-                fs.Spans.Add(new Span { Text = s });
-                label.FormattedText = fs;
-                layout.Children.Add(label);
-
-                var entry = new Entry { Placeholder = "Answer here..." };
-                layout.Children.Add(entry);
-            }
-
-			
-
-			var scrollView = new ScrollView { Content = layout };
-            this.Content = scrollView;
-
-			
 		}
 
-		async void OnClicked(object sender, EventArgs args)
+		async void OnClicked1(object sender, EventArgs args)
 		{
 			await Navigation.PushAsync(new Login.LoginPage());
 		}
 
-		private void Entry_Completed(object sender, EventArgs e)
-        {
-            String userEntry = ((Entry)sender).Text;
-            sendAndResponse(userEntry);
-        }
-        
-        private async void sendAndResponse(String userEntry)
-        {
-            using (var httpClient = new HttpClient())
-            {
-                simpleJsonClass sjc = new simpleJsonClass(userEntry);
-                var uri = new Uri("http://mizesolutions.com/a4aa1/a4aa/src/a4aa_rec.php");
-                var response = httpClient.PostAsync(uri, new StringContent(sjc.toJsonString(), Encoding.UTF8, "application/json")).Result;
-                resLabel.Text = await response.Content.ReadAsStringAsync();
-            }
-        }
-    }
+		async void OnClicked2(object sender, EventArgs args)
+		{
+			await Navigation.PushAsync(new SurveyPages.MainSurveyPage());
+		}
+	}
 }
