+++ HelloWorld/HelloWorld/SurveyClasses/SurveyTables/SeatingT.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace HelloWorld.SurveyClasses.SurveyTables
+namespace A4AA_Application.SurveyClasses.SurveyTables
 {
     class SeatingT
     {
